+++ main.py
@@ -3,7 +3,7 @@
 # Name: Drew McGregor
 # Class: CS30
 # Assignment: Capstone Coding Project
-# Version: 0.1
+# Version: 0.2
 #-----------------------------------------------------------------------------
 '''
    Here is the headers docString 
@@ -11,8 +11,11 @@
 '''
 #-----------------------------------------------------------------------------
 #-Imports and Global Variables------------------------------------------------
+import math as m
 import services as s
+import user as u
 #-Functions ------------------------------------------------------------------
 
 
 #-Main -----------------------------------------------------------------------
+print_list  = [s.Apple, s.Windows]