+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 Changelog
+v0.2----2024-5-31
+added user class
+added distance_from_user() function
+added Windows and Apple service objects
+added Bob user object
+
 v0.1
 added Service class __init__()
 added Service class save()